Fix most Mountain Fortress staircase bugs.
Previously, the staircase of the Mountain Fortress tended to burrow
underground and fail to stop. The code has now been changed so that
when the staircase reaches the ground, it terminates generation for
that chunk and labels it a 'dead chunk' for the remainder of that
player's session (or until it is removed by the addition of more
dead chunks). This ensures that no staircase generation will occur
in neighboring chunks, which generate independently.

This system is still under development and a bit buggy, although
less so than before!