Port to Minecraft 1.18.2.
Subterranea is now ported to MC 1.18.2! The code has been updated
accordingly to be compatible with the new version (all in structure
generation). The MIT License included in the last commit for
using TelepathicGrunt's StructureTutorialMod has been removed, as
the new structure generation system made the code borrowed from it
unnecessary (it has all been removed). However, TelepathicGrunt is
still mentioned in the credits for the vital help he and his mod
provided in understanding structure generation.

Also, the pack.mcmeta version number has been changed to 9 to
match the new Minecraft version! (1.18.2)