Reimplement Mountain Fortress generation.
The Mountain Fortress will now generate in the world again!
Most features of generation are now specified by data. Unfortunately,
there is a built-in structure depth limit of 7 pieces; a new codec
has been written for the Mountain Fortress to extend this to 15
pieces as the structure's recursion does not start for several pieces.